Release 0.90.2

+++ pytsite/file_storage_odm/_eh.py
@@ -16,7 +16,6 @@ def _update_0_90_0():
         db.get_collection('images').rename('file_images')
         msg = "Collection 'images' renamed to 'file_images'."
         logger.info(msg)
-        print(msg)
 
         images = db.get_collection('file_images')
         for doc in images.find():
@@ -27,7 +26,6 @@ def _update_0_90_0():
             })
             msg = 'Path updated for image: {}'.format(doc['_id'])
             logger.info(msg)
-            print(msg)
 
     # Move 'storage/image' to 'storage/file/image'
     images_dir = os.path.join(reg.get('paths.storage'), 'image')
@@ -35,20 +33,19 @@ def _update_0_90_0():
         files_dir = os.path.join(reg.get('paths.storage'), 'file')
 
         if not os.path.exists(files_dir):
-            os.makedirs(files_dir, 493)
+            os.makedirs(files_dir, 0o755)
 
         shutil.move(images_dir, files_dir)
 
         msg = '{} moved to {}.'.format(images_dir, files_dir)
         logger.info(msg)
-        print(msg)
 
-    # Delete 'static/image' directory
-    static_images_path = os.path.join(reg.get('paths.static'), 'image')
-    shutil.rmtree(static_images_path)
-    msg = '{} has been removed.'.format(static_images_path)
-    logger.info(msg)
-    print(msg)
+        # Delete 'static/image' directory
+        static_images_path = os.path.join(reg.get('paths.static'), 'image')
+        if os.path.exists(static_images_path):
+            shutil.rmtree(static_images_path)
+            msg = '{} has been removed.'.format(static_images_path)
+            logger.info(msg)
 
 
 def update(version: str):
